requirements.txt
argcomplete==3.6.2
asgiref==3.9.2
attrs==25.3.0
boto3==1.40.44
botocore==1.40.44
certifi==2025.8.3
cfn-flip==1.3.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
durationpy==0.10
hjson==3.1.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kappa==0.6.0
MarkupSafe==3.0.3
placebo==0.9.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
zappa==0.60.2
gunicorn